Saving of state when offline implemented
Added log saving and retrieving after shutdown (on the
.share/local/cleyto_coin/logs or something like that), added the config
toml (on .config/cleyto_coin), new binary for creating wallets and
sending transactions. Node still doesn't check the U... something to see
if the sender had enough to pay the receiver. Also changed how the
private and public keys are held in the wallet (from [u8] to the PKey
objects.